requirement.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
diffusers==0.30.2
Django==5.1
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.1
